\name{rhcclab-package}
\alias{rhcclab-package}
\alias{rhcclab}
\docType{package}
\title{
\packageTitle{rhcclab}
}
\description{
\packageDescription{rhcclab}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rhcclab}
\packageIndices{rhcclab}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rhcclab}

Maintainer: \packageMaintainer{rhcclab}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
